Tweak algorithm to rely on any substitution
The initial design focused on circular permutations as a state.
We are moving to allow any permutation, therefore letting each pass make
any random switch.

This allows the current slot in the state to be switched with any other
slot with equal probability, which intuitively should improve the
quality of the PRNG.